Closes #1585: Centralize yt back behaviour in new class
